Version 1.0.0

+++ insurance-policy-api-domain/Contracts/IInstallmentDomainService.cs
@@ -5,5 +5,5 @@ namespace insurance_policy_api_domain.Contracts;
 public interface IInstallmentDomainService
 {
     Task UpdateInstallmentiesAsync(IEnumerable<InstallmentEntity> installmentiesToUpdate);
-    Task RegisterPaymentAsync(int policyId, DateOnly datePayment);
+    Task RegisterPaymentAsync(long policyId, DateOnly datePayment);
 }
