+++ insurance-policy-api-domain/Contracts/IInstallmentRepository.cs
@@ -4,7 +4,7 @@ namespace insurance_policy_api_domain.Contracts;
 
 public interface IInstallmentRepository : IDisposable
 {
-    Task<InstallmentEntity> GetByIdAsync(int entityID);
+    Task<InstallmentEntity> GetByIdAsync(long entityID);
     Task UpdateAsync(InstallmentEntity installmentEntity);
     Task UpdateRangeAsync(IEnumerable<InstallmentEntity> installmentiesEntity);
 }
