+++ insurance-policy-api-domain/Contracts/IPolicyDomainService.cs
@@ -5,7 +5,7 @@ namespace insurance_policy_api_domain.Contracts;
 public interface IPolicyDomainService
 {
     Task CreateNewPolicyAsync(PolicyEntity policyEntity);
-    Task<PolicyEntity> RetrievePolicyByIdAsync(int policyId);
+    Task<PolicyEntity> RetrievePolicyByIdAsync(long policyId);
     Task<IEnumerable<PolicyEntity>> RetrieveAllPoliciesAsync(int skip, int take);
     Task UpdatePolicyAsync(PolicyEntity policyEntity);
 }
