+++ insurance-policy-api-domain/Contracts/IPolicyRepository.cs
@@ -5,7 +5,7 @@ namespace insurance_policy_api_domain.Contracts;
 public interface IPolicyRepository : IDisposable
 {
     Task AddAsync(PolicyEntity policyEntity);
-    Task<PolicyEntity> GetByIdAsync(int entityID);
+    Task<PolicyEntity> GetByIdAsync(long entityID);
     Task<IEnumerable<PolicyEntity>> GetAllAsync(int skip, int take);
     Task UpdateAsync(PolicyEntity policyEntity);
