+++ insurance-policy-api/Interfaces/IPolicyAppService.cs
@@ -5,7 +5,7 @@ namespace insurance_policy_api.Interfaces;
 public interface IPolicyAppService
 {
     Task<PolicyDTO> CreatePolicyAsync(PolicyDTO policyDto);
-    Task<PolicyDTO> GetPolicyByIdAsync(int entityId);
+    Task<PolicyDTO> GetPolicyByIdAsync(long entityId);
     Task<IEnumerable<PolicyDetailsDTO>> GetAllPoliciesAsync(int skip, int take);
     Task<PolicyDTO> UpdatePolicyAsync(PolicyDTO policyDto);
 }
